+++ zarrs/src/array/codec/array_to_bytes/sharding/sharding_partial_decoder_sync.rs
@@ -6,7 +6,10 @@ use rayon::prelude::*;
 use unsafe_cell_slice::UnsafeCellSlice;
 use zarrs_chunk_grid::{ArraySubset, ChunkGridTraits};
 
-use super::{ShardingCodecOptions, ShardingIndexLocation, calculate_chunks_per_shard};
+use super::{
+    ShardingCodecOptions, ShardingIndexLocation, calculate_chunks_per_shard,
+    nested_local_subchunk_grids,
+};
 use crate::array::array_bytes_internal::merge_chunks_vlen;
 use crate::array::chunk_grid::RegularChunkGrid;
 use crate::array::{
@@ -218,15 +221,16 @@ impl ArrayPartialDecoderTraits for ShardingPartialDecoder {
         )
     }
 
-    fn local_subchunk_grid(
+    fn local_subchunk_grids(
         &self,
         _options: &CodecOptions,
-    ) -> Result<Option<ChunkGrid>, CodecError> {
+    ) -> Result<Vec<Option<ChunkGrid>>, CodecError> {
         let shard_shape = bytemuck::must_cast_slice(&self.shard_shape).to_vec();
-        Ok(Some(ChunkGrid::new(
+        let subchunk_grid = ChunkGrid::new(
             RegularChunkGrid::new(shard_shape, self.subchunk_shape.clone())
                 .map_err(|err| CodecError::Other(err.to_string()))?,
-        )))
+        );
+        nested_local_subchunk_grids(subchunk_grid, &self.inner_codecs)
     }
 
     fn partial_decode_into(